+++ modules/Contact/Database/Migrations/2020_02_07_182535_create_contact_details_table.php
@@ -16,12 +16,12 @@ class CreateContactDetailsTable extends Migration
         Schema::create(config('crmomni-migration.table_name.contact.details'), function (Blueprint $table) {
             $table->bigIncrements('id');
 
-            $table->integer('org_id');
+            $table->unsignedBigInteger('org_id');
             $table->integer('contact_id');
-            $table->integer('type_id')->nullable();
+            $table->unsignedBigInteger('type_id')->nullable();
             $table->integer('subtype_id')->nullable();
 
-            $table->integer('country_id')->nullable();
+            $table->unsignedBigInteger('country_id')->nullable();
             $table->string('identifier')->nullable();
             $table->string('proxy')->nullable();
 
@@ -30,8 +30,8 @@ class CreateContactDetailsTable extends Migration
             $table->boolean('is_active')->default(true);
 
             //Audit Log Fields
-            $table->integer('created_by')->default(0);
-            $table->integer('updated_by')->nullable();
+            $table->unsignedBigInteger('created_by')->default(0);
+            $table->unsignedBigInteger('updated_by')->nullable();
             $table->integer('deleted_by')->nullable();
             
             $table->timestamps();
