+++ modules/Contact/Database/Migrations/2020_02_07_182545_create_society_address_table.php
@@ -19,8 +19,8 @@ class CreateSocietyAddressTable extends Migration
             $table->string('name');        
             $table->string('locality')->nullable();
             $table->string('city')->nullable();
-            $table->integer('state_id')->nullable();
-            $table->integer('country_id')->nullable();
+            $table->unsignedBigInteger('state_id')->nullable();
+            $table->unsignedBigInteger('country_id')->nullable();
             $table->string('zipcode')->nullable();
 
             //Location Details
@@ -31,8 +31,8 @@ class CreateSocietyAddressTable extends Migration
             $table->boolean('is_active')->default(true);
 
             //Audit Log Fields
-            $table->integer('created_by')->default(0);
-            $table->integer('updated_by')->nullable();
+            $table->unsignedBigInteger('created_by')->default(0);
+            $table->unsignedBigInteger('updated_by')->nullable();
             $table->integer('deleted_by')->nullable();
             
             $table->timestamps();
