+++ modules/Contact/Database/Migrations/2020_02_07_182547_create_apartment_address_table.php
@@ -20,8 +20,8 @@ class CreateApartmentAddressTable extends Migration
             $table->integer('society_id')->nullable();            
             $table->string('locality')->nullable();
             $table->string('city')->nullable();
-            $table->integer('state_id')->nullable();
-            $table->integer('country_id')->nullable();
+            $table->unsignedBigInteger('state_id')->nullable();
+            $table->unsignedBigInteger('country_id')->nullable();
             $table->string('zipcode')->nullable();
 
             //Location Details
@@ -32,8 +32,8 @@ class CreateApartmentAddressTable extends Migration
             $table->boolean('is_active')->default(true);
 
             //Audit Log Fields
-            $table->integer('created_by')->default(0);
-            $table->integer('updated_by')->nullable();
+            $table->unsignedBigInteger('created_by')->default(0);
+            $table->unsignedBigInteger('updated_by')->nullable();
             $table->integer('deleted_by')->nullable();
             
             $table->timestamps();
