+++ modules/Core/Database/Migrations/2020_02_05_182201_create_currency_table.php
@@ -28,8 +28,8 @@ class CreateCurrencyTable extends Migration
             $table->boolean('is_active')->default(true);
 
             //Audit Log Fields
-            $table->integer('created_by')->default(0);
-            $table->integer('updated_by')->nullable();
+            $table->unsignedBigInteger('created_by')->default(0);
+            $table->unsignedBigInteger('updated_by')->nullable();
 
             $table->timestamps();
         });
