+++ modules/Core/Database/Migrations/2020_02_05_182423_create_role_privileges_table.php
@@ -21,8 +21,8 @@ class CreateRolePrivilegesTable extends Migration
             $table->boolean('is_active')->default(true);
 
             //Audit Log Fields
-            $table->integer('created_by')->default(0);
-            $table->integer('updated_by')->nullable();
+            $table->unsignedBigInteger('created_by')->default(0);
+            $table->unsignedBigInteger('updated_by')->nullable();
 
             $table->timestamps();
         });
