+++ modules/Document/Database/Migrations/2020_02_05_182555_create_documents_table.php
@@ -17,7 +17,7 @@ class CreateDocumentsTable extends Migration
             $table->bigIncrements('id');
 
             // FK Relationships
-            $table->integer('org_id');
+            $table->unsignedBigInteger('org_id');
             $table->integer('entity_type_id');
             $table->integer('reference_id');
 
@@ -30,8 +30,8 @@ class CreateDocumentsTable extends Migration
             $table->boolean('is_active')->default(true);
 
             //Audit Log Fields
-            $table->integer('created_by')->default(0);
-            $table->integer('updated_by')->nullable();
+            $table->unsignedBigInteger('created_by')->default(0);
+            $table->unsignedBigInteger('updated_by')->nullable();
             $table->integer('deleted_by')->nullable();
 
             $table->timestamps();
