+++ modules/Subscription/Database/Migrations/2020_07_27_093848_create_subscriptionplan_table.php
@@ -26,8 +26,8 @@ class CreateSubscriptionplanTable extends Migration
             $table->boolean('is_active')->default(true);
 
             //Audit Log Fields
-            $table->integer('created_by')->default(0);
-            $table->integer('updated_by')->nullable();
+            $table->unsignedBigInteger('created_by')->default(0);
+            $table->unsignedBigInteger('updated_by')->nullable();
             $table->integer('deleted_by')->nullable();
 
             $table->timestamps();
